Refactor spotlight data fetching into separate utility functions
