+++ pages/index.js
@@ -9,147 +9,16 @@ import { ThreeCampuses } from '@/components/home/three-campuses';
 import { SpotlightHero } from '@/components/home/spotlight-hero';
 import { HomeStory } from '@/components/home/story';
 import { StudyHere } from '@/components/home/study-here';
+import { getSpotlightCards, getSpotlightHero } from '@/data/drupal/home';
 
 export async function getStaticProps(context) {
-	const isPreview = context?.preview || process.env.NODE_ENV !== 'production';
-	const status = isPreview ? null : true;
-
-	// Look through the latest revisions of spotlight items, filtered to only those that are rank 1 (and published if not in preview mode).
-	const hero = (
-		await graphql(
-			`
-				query SpotlightHero($status: Boolean = null) {
-					spotlightRevisions(pageSize: 1, filter: { rank: ["1"], status: $status }) {
-						results {
-							... on NodeSpotlight {
-								id
-								caption
-								captionAlignment
-								thumbnailImageCrop
-								title
-								url {
-									url
-									title
-								}
-								image {
-									... on MediaImage {
-										image {
-											height
-											alt
-											width
-											variations(styles: SPOTLIGHT_HERO_IMAGE) {
-												url
-											}
-										}
-									}
-								}
-							}
-						}
-					}
-				}
-			`,
-			{
-				status: status,
-			},
-		)
-	)?.data?.spotlightRevisions?.results?.[0];
-
-	// For spotlight cards our query is a bit more involved. We look through the latest revisions (5 at once this time) like we did for the hero, but because we need 4 cards, and a single spotlight can have many revisions, we keep track of the ids of the spotlights we've seen and keep requesting more until we have 4 unique ids. With those ids. we can request the content separately.
-
-	const ids = new Set();
-	let isLastPage = false;
-	let page = 0;
-
-	outer: while (!isLastPage) {
-		const cards = (
-			await graphql(
-				`
-					query SpotlightCardIDs($status: Boolean = null, $page: Int = 0) {
-						spotlightRevisions(filter: { rank: [], status: $status }, page: $page, pageSize: 5) {
-							results {
-								... on NodeSpotlight {
-									id
-								}
-							}
-						}
-					}
-				`,
-				{
-					status: status,
-					page: page,
-				},
-			)
-		)?.data?.spotlightRevisions?.results;
-
-		for (const card of cards) {
-			if (card.id === hero.id) {
-				continue;
-			}
-
-			if (ids.size === 4) {
-				break outer;
-			}
-
-			ids.add(card.id);
-		}
-
-		// The page wasn't full, meaning we were on the last page.
-		if (cards.length < 5) {
-			isLastPage = true;
-		}
-
-		page++;
-	}
-
-	// Now that we have the ids, we can retrieve the content.
-	const cards = await Promise.all(
-		Array.from(ids).map(async (id) => {
-			const { data } = await graphql(
-				`
-					query SpotlightCard($status: Boolean = false, $id: String = "") {
-						spotlightRevisions(filter: { status: $status, id: $id }, pageSize: 1) {
-							results {
-								... on NodeSpotlight {
-									image {
-										... on MediaImage {
-											image {
-												alt
-												height
-												width
-												variations(styles: SPOTLIGHT_CARD_IMAGE) {
-													url
-												}
-											}
-										}
-									}
-									rank
-									title
-									thumbnailImageCrop
-									url {
-										url
-										title
-									}
-								}
-							}
-						}
-					}
-				`,
-				{
-					status: status,
-					id: id,
-				},
-			);
-
-			return {
-				id: id,
-				...(data.spotlightRevisions.results?.[0] ?? {}),
-			};
-		}),
-	);
+	const status = context?.preview || process.env.NODE_ENV !== 'production' ? null : true;
+	const hero = await getSpotlightHero(status);
+	const cards = await getSpotlightCards(status, hero);
 
 	return {
 		props: {
-			cards: cards.sort((a, b) => a.rank - b.rank),
+			cards: cards,
 			hero: hero,
 		},
 	};
@@ -164,22 +33,30 @@ export default function Home({ cards, hero }) {
 				<TagLine />
 
 				<div className="pt-1">
-					<Heading level={2}>Our Latest News and Events</Heading>
+					<Heading level={2} className="font-condensed">
+						Our Latest News and Events
+					</Heading>
 					<SpotlightCards cards={cards} />
 				</div>
 
 				<div className="pt-1">
-					<Heading level={2}>Study Here</Heading>
+					<Heading level={2} className="font-condensed">
+						Study Here
+					</Heading>
 					<StudyHere />
 				</div>
 
 				<div className="pt-1">
-					<Heading level={2}>How We Rank Among the World</Heading>
+					<Heading level={2} className="font-condensed">
+						How We Rank Among the World
+					</Heading>
 					<Rankings />
 				</div>
 
 				<div className="pt-1">
-					<Heading level={2}>Our Three Campuses</Heading>
+					<Heading level={2} className="font-condensed">
+						Our Three Campuses
+					</Heading>
 					<ThreeCampuses />
 				</div>
 			</Container>
